requirements.txt
aio-pika==9.6.1
aiokafka==0.13.0
aiormq==6.9.3
aiosqlite==0.22.1
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
bandit==1.9.4
bcrypt==4.0.1
billiard==4.2.4
bleach==6.3.0
boto3==1.42.69
botocore==1.42.69
cachetools==7.0.5
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
chaostoolkit==1.19.0
chaostoolkit-lib==1.44.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.4
cryptography==46.0.5
Deprecated==1.3.1
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
elastic-transport==9.2.1
elasticsearch==9.3.0
email-validator==2.3.0
Faker==40.11.0
fakeredis==2.34.1
fastapi==0.135.1
filelock==3.25.0
flake8==7.3.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hvac==2.4.0
identify==2.6.17
idna==3.11
importlib_metadata==9.0.0
iniconfig==2.3.0
jmespath==1.1.0
kombu==5.6.2
limits==5.8.0
line-profiler==4.1.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
multidict==6.7.1
nodeenv==1.10.0
packaging==26.0
pact-python==3.2.1
pact-python-cli==2.5.7.0
pact-python-ffi==0.4.28.2
pamqp==3.3.0
passlib==1.7.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
py-spy==0.3.14
pyasn1==0.6.2
pybreaker==1.4.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyinstrument==4.6.2
pylogbeat==2.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-consul2==0.1.5
python-dateutil==2.9.0.post0
python-discovery==1.1.1
python-dotenv==1.2.2
python-jose==3.5.0
python-json-logger==4.0.0
python-logstash-async==4.1.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rich==14.3.3
rsa==4.9.1
ruff==0.15.5
s3transfer==0.16.0
scalar_fastapi==1.8.0
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
starlette==0.52.1
stevedore==5.7.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vine==5.1.0
virtualenv==21.1.0
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.0
